# docs/swagger.yaml
basePath: /
definitions:
...
  model.Sesion:
    properties:
      comentarios:
        type: string
      duracion_min:
        type: integer
      fecha:
        type: string
      id:
        type: integer
      usuario_id:
        type: integer
    type: object
...
  model.TipoEjercicio:
    properties:
      id:
        type: integer
      nombre:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestión de rutinas de gimnasio
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GymBro API
  version: "1.0"
paths:
...
  /sesion:
    get:
      consumes:
      - application/json
      description: Obtiene la lista completa de sesiones de entrenamiento
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Sesion'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener todas las sesiones
      tags:
      - sesiones
    post:
      consumes:
      - application/json
      description: Crea una nueva sesión de entrenamiento en el sistema
      parameters:
      - description: Datos de la sesión
        in: body
        name: sesion
        required: true
        schema:
          $ref: '#/definitions/model.Sesion'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Sesion'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear nueva sesión
      tags:
      - sesiones
...
  /sesion/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una sesión del sistema
      parameters:
      - description: ID de la sesión
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Sesión eliminada
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar sesión
      tags:
      - sesiones
    get:
      consumes:
      - application/json
      description: Obtiene una sesión específica por su ID
      parameters:
      - description: ID de la sesión
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Sesion'
        "404":
          description: Sesión no encontrada
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener sesión por ID
      tags:
      - sesiones
    put:
      consumes:
      - application/json
      description: Actualiza una sesión existente
      parameters:
      - description: ID de la sesión
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados de la sesión
        in: body
        name: sesion
        required: true
        schema:
          $ref: '#/definitions/model.Sesion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Sesion'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar sesión
      tags:
      - sesiones
  /sesion/fecha:
    get:
      consumes:
      - application/json
      description: Obtiene todas las sesiones dentro de un rango de fechas específico
      parameters:
      - description: 'Fecha desde (formato ISO: 2024-01-02T15:04:05Z)'
        in: query
        name: fechaDesde
        required: true
        type: string
      - description: 'Fecha hasta (formato ISO: 2024-01-02T15:04:05Z)'
        in: query
        name: fechaHasta
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Sesion'
            type: array
        "400":
          description: Formato de fecha inválido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener sesiones por rango de fechas
      tags:
      - sesiones
  /sesion/usuario/{id}:
    get:
      consumes:
      - application/json
      description: Obtiene todas las sesiones de un usuario específico
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Sesion'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener sesiones por usuario
      tags:
      - sesiones
  /tipo-ejercicio:
    get:
      consumes:
      - application/json
      description: Obtiene la lista completa de tipos de ejercicio
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TipoEjercicio'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener todos los tipos de ejercicio
      tags:
      - tipos-ejercicio
    post:
      consumes:
      - application/json
      description: Crea un nuevo tipo de ejercicio en el sistema
      parameters:
      - description: Datos del tipo de ejercicio
        in: body
        name: tipo
        required: true
        schema:
          $ref: '#/definitions/model.TipoEjercicio'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TipoEjercicio'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear nuevo tipo de ejercicio
      tags:
      - tipos-ejercicio
  /tipo-ejercicio/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un tipo de ejercicio del sistema
      parameters:
      - description: ID del tipo de ejercicio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Tipo de ejercicio eliminado
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar tipo de ejercicio
      tags:
      - tipos-ejercicio
    get:
      consumes:
      - application/json
      description: Obtiene un tipo de ejercicio específico por su ID
      parameters:
      - description: ID del tipo de ejercicio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TipoEjercicio'
        "404":
          description: Tipo de ejercicio no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener tipo de ejercicio por ID
      tags:
      - tipos-ejercicio
    put:
      consumes:
      - application/json
      description: Actualiza un tipo de ejercicio existente
      parameters:
      - description: ID del tipo de ejercicio
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados del tipo de ejercicio
        in: body
        name: tipo
        required: true
        schema:
          $ref: '#/definitions/model.TipoEjercicio'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TipoEjercicio'
        "400":
          description: Datos inválidos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Tipo de ejercicio no encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar tipo de ejercicio
      tags:
      - tipos-ejercicio
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"